scrapy/scraping_service/infrastructure/scraping/mercado_libre_spider.py
"""
Capa de infraestructura - Mercado Libre Spider
Spider especifico para Mercado Libre Mexico
"""
from typing import List
from decimal import Decimal
from urllib.parse import quote_plus
import httpx
from bs4 import BeautifulSoup

from .base_spider import BaseSpider
from ...domain.product_offer import ProductOffer
from ...domain.search_params import SearchParams
from ...domain.provider_enum import Provider
from ...domain.errors import ScrapingError


class MercadoLibreSpider(BaseSpider):
    """
    Spider para Mercado Libre Mexico
    
    URL base: https://www.mercadolibre.com.mx
    Busqueda: https://listado.mercadolibre.com.mx/{query}
    """
    
    def __init__(self):
        super().__init__(Provider.MERCADO_LIBRE)
    
    def build_search_url(self, search_params: SearchParams) -> str:
        """
        Construye URL de busqueda para ML
        
        Ejemplo: https://listado.mercadolibre.com.mx/bujia-ngk-honda-civic-2015
        """
        query = search_params.to_query_string()
        encoded_query = quote_plus(query)
        return f"https://listado.mercadolibre.com.mx/{encoded_query}"
    
    async def scrape(self, search_params: SearchParams) -> List[ProductOffer]:
        """
        Realiza scraping en Mercado Libre
        
        TODO: Implementar parseo real con BeautifulSoup/Scrapy
        """
        try:
            url = self.build_search_url(search_params)
            
            # TODO: Realizar request HTTP y parsear resultados
            # Por ahora retornar lista vacia
            products = []
            
            # Ejemplo de implementacion (PLACEHOLDER):
            # async with httpx.AsyncClient() as client:
            #     response = await client.get(url, headers=self._get_headers())
            #     soup = BeautifulSoup(response.text, 'html.parser')
            #     products = self._parse_products(soup)
            
            print(f"  Scraping ML: {url}")
            return products
            
        except Exception as e:
            raise ScrapingError(
                provider=self.provider.value,
                message=f"Error en scraping: {str(e)}"
            )
    
    def _get_headers(self) -> dict:
        """Headers para requests HTTP"""
        return {
            "User-Agent": "Mozilla/5.0 (Windows NT 10.0; Win64; x64) AppleWebKit/537.36",
            "Accept": "text/html,application/xhtml+xml,application/xml;q=0.9,*/*;q=0.8",
            "Accept-Language": "es-MX,es;q=0.9,en;q=0.8"
        }
    
    def _parse_products(self, soup: BeautifulSoup) -> List[ProductOffer]:
        """
        Parsea HTML y extrae productos
        
        TODO: Implementar selectores CSS reales de ML
        """
        products = []
        
        # PLACEHOLDER - Selectores de ejemplo (ajustar a HTML real)
        # items = soup.select('.ui-search-result__content')
        # 
        # for item in items:
        #     titulo = item.select_one('.ui-search-item__title').text.strip()
        #     precio_text = item.select_one('.price-tag-amount').text
        #     precio = self._parse_price(precio_text)
        #     url = item.select_one('a')['href']
        #     
        #     product = ProductOffer(
        #         titulo=titulo,
        #         precio=precio,
        #         url=url,
        #         fuente=self.provider,
        #         metadata={}
        #     )
        #     products.append(product)
        
        return products
    
    def _parse_price(self, price_text: str) -> Decimal:
        """
        Parsea texto de precio a Decimal
        
        Ejemplo: "$1,250" -> Decimal("1250.00")
        """
        # Remover simbolos y comas
        clean_price = price_text.replace('$', '').replace(',', '').strip()
        return Decimal(clean_price)
