scrapy/requirements.txt
# Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0

# Scraping
scrapy==2.11.0
scrapy-playwright==0.0.34  # Para sitios con JavaScript
playwright==1.40.0

# Database
asyncpg==0.29.0
psycopg2-binary==2.9.9
sqlalchemy==2.0.25

# HTTP Client
httpx==0.26.0
aiohttp==3.9.1

# Utilities
python-dotenv==1.0.0
python-dateutil==2.8.2
pytz==2023.3

# Data Processing
lxml==5.0.0
beautifulsoup4==4.12.2

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
pytest-mock==3.12.0

# Logging
loguru==0.7.2

# Cache
redis==5.0.1  # Opcional: para cache distribuido

# Validation
validators==0.22.0

# Development
black==23.12.1
flake8==7.0.0
mypy==1.8.0
